Fixed Photo_Screen_Saver crashes when router dies brutally. Fixed warnings.

+++ src/Photo_Screen_Saver/ZoomBase.cpp
@@ -16,23 +16,23 @@ ZoomBase::ZoomBase(Frame* PictureObject,
 
 	if(ScreenWidth < Width)
 	{
-		this->Height = Height*(float)ScreenWidth/Width + 1;
+		this->Height = int(Height*(float)ScreenWidth/Width + 1);
 		this->Width = ScreenWidth;
 	}
 	if(ScreenHeight < Height)
 	{
-		this->Width = Width*(float)ScreenWidth/Width + 1;
+		this->Width = int(Width*(float)ScreenWidth/Width + 1);
 		this->Height = ScreenHeight;
 	}
 
 	if(ScreenWidth > Width)
 	{
-		this->Height = Height*(float)ScreenWidth/Width + 1;
+		this->Height = int(Height*(float)ScreenWidth/Width + 1);
 		this->Width = ScreenWidth;
 	}
 	if(ScreenHeight > Height)
 	{
-		this->Width = Width*(float)ScreenWidth/Width + 1;
+		this->Width = int(Width*(float)ScreenWidth/Width + 1);
 		this->Height = ScreenHeight;
 	}
 }
